Everything works. Huge performance gains.
The points in the data are still sorted by angle relative to the
centroid, but they are selected through the GA by index into that list,
rather than angle. Execution time for identify.py on sample2.png is down
from ~30 seconds to ~1 second, with the final fitness being just as good
(around 2100).