Added an option to select fitness function residue calculation method.
residue_method, an input to the PolygonTester, can be set either to
'line' or 'segment'. When set to 'line', the residue for each individual
is calculated as part of the orthogonal regression and represents the
sum of the orthogonal distance from each point to the infinite extension
of its particular line (the old behavior). When set to 'segment', it
instead calculates the distance from the point to the nearest point on
its respective line segment. Thus, a point which lies on the line but
not on the line segment will have a non-zero residual in this case.

I haven't tested this yet.